requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.61.1
requests==2.31.0
gunicorn==21.2.0
pdfplumber==0.10.2
python-magic==0.4.27
